requirements.txt
pip
pygame-ce
